+++ bgpd/fcserver/include/fcserver.h
@@ -30,9 +30,7 @@
 #include "libncs6.h"
 #include "defines.h"
 
-#define FC_DEFAULT_CONFIG_FNAME "/etc/frr/assets/config.json"
 #define FC_PROGRAM_NAME "fcserver"
-#define FC_PORT 23160
 #define FC_EPOLL_MAX_EVENTS 32
 #define FC_SOCK_BACKLOG 32